back/api/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
PyJWT==2.1.0
pymongo==4.1.1
